third_party/blink/web_tests/external/wpt/css/css-display/parsing/display-valid-expected.txt

This is a testharness.js-based test.
Found 28 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] e.style['display'] = "run-in" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['display'] = "ruby-base" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['display'] = "run-in flow" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['display'] = "flow run-in" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['display'] = "flow-root run-in" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['display'] = "run-in flow-root" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['display'] = "flex run-in" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['display'] = "run-in flex" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['display'] = "grid run-in" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['display'] = "run-in grid" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['display'] = "table run-in" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['display'] = "run-in table" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['display'] = "run-in ruby" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['display'] = "ruby run-in" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['display'] = "run-in list-item" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['display'] = "list-item run-in" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['display'] = "flow run-in list-item" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['display'] = "run-in flow list-item" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['display'] = "flow list-item run-in" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['display'] = "run-in list-item flow" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['display'] = "list-item run-in flow" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['display'] = "list-item flow run-in" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['display'] = "flow-root run-in list-item" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['display'] = "run-in flow-root list-item" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['display'] = "flow-root list-item run-in" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['display'] = "run-in list-item flow-root" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['display'] = "list-item run-in flow-root" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['display'] = "list-item flow-root run-in" should set the property value
  assert_not_equals: property should be set got disallowed value ""
Harness: the test ran to completion.
